Move `ConsoleLogger` under logger package
