+++ yaml_requests/logger/__init__.py
@@ -1,0 +1,1 @@
+from ._console import ConsoleLogger
